frontend-ux/requirements.txt
streamlit>=1.33
google-cloud-firestore
requests
plotly
sseclient-py
Pillow
